requirements.txt
fastap
uvicorn
prometheusrock
starlette
botocore
boto3

pytest
coverage
httpx